Unified the stream classes (#1332)
* Unified the DataSourceStreams to be based on Stream interface.
* Implemented long standing refactoring issue.
* Small bug fixes.
* Deprecate `length()` method before removing it completely.

+++ Sming/Libraries/ArduCAM/ArduCAMStream.h
@@ -40,7 +40,7 @@ public:
 	}
 
 	bool dataReady();
-	size_t available();
+	int available();
 
 private:
